requirements.txt
pybit>=5.7.0
python-dotenv>=1.0.0
click>=8.1.0
tabulate>=0.9.0
requests>=2.31.0
psycopg2-binary>=2.9.0
